ivcanon screens out global vars as IV
Suppressed an overflow assertion with debug compiler in me_ivopts.cpp

+++ src/mapleall/maple_me/src/me_ivopts.cpp
@@ -1432,7 +1432,7 @@ MeExpr *IVOptimizer::ComputeExtraExprOfBase(MeExpr &candBase, MeExpr &groupBase,
   for (auto &itCand : candMap) {
     auto itGroup = groupMap.find(itCand.first);
     if (itCand.first == kInvalidExprID) {
-      candConst = itCand.second.second * ratio;
+      candConst = (uint64)itCand.second.second * ratio;
       groupConst = itGroup == groupMap.end() ? 0 : itGroup->second.second;
       continue;
     }
